Isolate parsing functions that were in prec.js
